Add destructor for Integration Test Framework
According to the source code, there is no need to ensure that done() would not be called twice. Since it is safe to call done() twice, neither done() nor ~() does not have check for multiple calls.

If it will turn out that done() cannot be safely called twice, additional commit will be issued.

Signed-off-by: Igor Egorov <igor@soramitsu.co.jp>